Include statements for API headers became consistent.
It is fine when header files can be referenced in include statements without
the specification of a directory. But a few file names were chosen for C++ source
files here that have got the potential to get in conflict with name selections
from other software libraries. The name "libsocket" was therefore added as
a parent directory to some includes in a consistent way.

The affected source code places were also adjusted so that angle brackets are
used there now.
http://stackoverflow.com/questions/21593/what-is-the-difference-between-include-filename-and-include-filename

Signed-off-by: Markus Elfring <elfring@users.sourceforge.net>