Fix (expected) buffering bugs in dgram_over_stream.
Due to the wrong use of do { } while () as well as the wrong counters,
the dgram_over_stream wrapper would block in certain conditions.

Namely, if the frame length is longer than the internal buffer, it would
first try to receive too many bytes, and then block while trying to
receive the non-existing rest of the frame.